Create new configuration storage

+++ ZenPlatform.EntityComponent/Migrations/SingleEntityMigrator.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using MoreLinq.Extensions;
+using ZenPlatform.Configuration.CompareTypes;
 using ZenPlatform.Configuration.Contracts;
 using ZenPlatform.Configuration.Data.Contracts.Entity;
 using ZenPlatform.Configuration.Structure.Data.Types.Complex;
@@ -35,10 +36,9 @@ namespace ZenPlatform.EntityComponent.Migrations
         }
 
 
-        public SSyntaxNode GetStep1(IXCObjectType oldBase, IXCObjectType actualBase, DDLQuery query)
+        public SSyntaxNode GetStep1(IXCObjectType old, IXCObjectType actual, DDLQuery query)
         {
-            XCSingleEntity old = (XCSingleEntity)oldBase;
-            XCSingleEntity actual = (XCSingleEntity)actualBase;
+             
 
             if (old == null && actual == null)
             {
@@ -52,7 +52,7 @@ namespace ZenPlatform.EntityComponent.Migrations
             {
                 var tableBuilder = query.Create().Table(actual.RelTableName);
 
-                foreach (var property in actual.Properties)
+                foreach (var property in actual.GetProperties())
                 {
                     property.GetPropertySchemas().ForEach(s =>
                     {
@@ -63,20 +63,27 @@ namespace ZenPlatform.EntityComponent.Migrations
                 
             }
             else
-            query.Copy().Table().FromTable(old.RelTableName).ToTable($"{actual.RelTableName}_tmp");
+            {
+                var comparer = new XCObjectTypeComparer<IXCObjectType>();
+                if (!comparer.Equals(old, actual))
+                {
+                    query.Copy().Table().FromTable(old.RelTableName).ToTable($"{actual.RelTableName}_tmp");
+                }
+            }
+            
 
             return query.Expression;
         }
 
-        public SSyntaxNode GetStep2(IXCObjectType oldBase, IXCObjectType actualBase, DDLQuery query)
+        public SSyntaxNode GetStep2(IXCObjectType old, IXCObjectType actual, DDLQuery query)
         {
-            XCSingleEntity old = (XCSingleEntity)oldBase;
-            XCSingleEntity actual = (XCSingleEntity)actualBase;
 
 
             if (old != null && actual != null)
             {
 
+
+                var comparer = new XCObjectTypeComparer<IXCObjectProperty>();
                 string tableName = $"{actual.RelTableName}_tmp";
 
                 var props = old.GetProperties()
@@ -97,7 +104,8 @@ namespace ZenPlatform.EntityComponent.Migrations
                         DeleteProperty(query, property.old, tableName);
                     } else
                     {
-                        ChangeProperty(query, property.old, property.actual, tableName);
+                        if (!comparer.Equals(property.old, property.actual))
+                            ChangeProperty(query, property.old, property.actual, tableName);
                     }
                 }
                 
@@ -164,24 +172,20 @@ namespace ZenPlatform.EntityComponent.Migrations
         }
 
 
-        public SSyntaxNode GetStep3(IXCObjectType oldBase, IXCObjectType actualBase, DDLQuery query)
+        public SSyntaxNode GetStep3(IXCObjectType old, IXCObjectType actual, DDLQuery query)
         {
-            XCSingleEntity old = (XCSingleEntity)oldBase;
-            XCSingleEntity actual = (XCSingleEntity)actualBase;
-
-            if (old != null && actual != null)
+            var comparer = new XCObjectTypeComparer<IXCObjectType>();
+            if (old != null && actual != null && !comparer.Equals(old, actual))
             {
                 query.Delete().Table($"{actual.RelTableName}");
             }
             return query.Expression;
         }
 
-        public SSyntaxNode GetStep4(IXCObjectType oldBase, IXCObjectType actualBase, DDLQuery query)
+        public SSyntaxNode GetStep4(IXCObjectType old, IXCObjectType actual, DDLQuery query)
         {
-            XCSingleEntity old = (XCSingleEntity)oldBase;
-            XCSingleEntity actual = (XCSingleEntity)actualBase;
-
-            if (old != null && actual != null)
+            var comparer = new XCObjectTypeComparer<IXCObjectType>();
+            if (old != null && actual != null && !comparer.Equals(old,actual))
             {
                 query.Rename().Table($"{actual.RelTableName}_tmp").To(old.RelTableName);
             }
